requirements.txt
amqp==2.4.2
billiard==3.5.0.5
celery==4.2.1
certifi==2019.3.9
chardet==3.0.4
dataclasses==0.6
Django==2.1.7
django-filter==2.1.0
django-stubs==0.9.0
djangorestframework==3.9.2
djangorestframework-stubs==0.3.0
Faker==0.9.1
idna==2.8
kombu==4.4.0
mixer==6.1.3
mypy==0.670
mypy-extensions==0.4.1
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
six==1.12.0
text-unidecode==1.2
typed-ast==1.3.1
typing-extensions==3.7.2
urllib3==1.24.1
vine==1.2.0
